more reworks and moves. load admin area details
